+++ Visual Studio Projects/ZWaveJS.NET/Demo Application/MainPage.xaml.cs
@@ -1,4 +1,6 @@
-﻿namespace Demo_Application;
+﻿using CommunityToolkit.Mvvm.ComponentModel;
+
+namespace Demo_Application;
 
 public partial class MainPage : ContentPage
 {
@@ -7,6 +9,7 @@ public partial class MainPage : ContentPage
 	public MainPage()
 	{
 		InitializeComponent();
+		BindingContext = new ConnectViewModel();
 	}
 
 	private void ConnectClick(object? sender, EventArgs e)
@@ -16,7 +19,7 @@ public partial class MainPage : ContentPage
 		switch (PK_Mode.SelectedItem.ToString())
 		{
 			case "Host":
-			 App._Instance.StartDriver(true,TX_Target.Text);
+			 App._Instance.StartDriver(true,PK_Port.SelectedItem.ToString()!);
 			break;
 
 			case "Client":
@@ -24,6 +27,32 @@ public partial class MainPage : ContentPage
 			break;
 		}
 		
-		
 	}
 }
+
+public partial class ConnectViewModel : ObservableObject
+{
+	public ConnectViewModel()
+	{
+		Ports = ZWaveJS.NET.Lib.SerialPorts();
+		Mode = "Select One";
+	}
+
+	partial void OnModeChanged(string value)
+	{
+		OnPropertyChanged(nameof(IsHost));
+		OnPropertyChanged(nameof(IsClient));
+	}
+
+	public string[] Ports {get;}
+
+	[ObservableProperty]
+	public string mode;
+
+	public bool IsHost => Mode == "Host";
+	public bool IsClient => Mode == "Client";
+
+	
+}
+
+